rename Model file to NexusModel
